requirements.txt
flask==3.0.0
numpy==1.24.3
tensorflow==2.11.0
matplotlib
seaborn
pandas
dynamo-release==1.3.0
joblib